+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/HyCollection.java
@@ -1,5 +1,6 @@
 package com.ruoyi.system.domain;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import org.apache.commons.lang3.builder.ToStringBuilder;
@@ -41,7 +42,7 @@ public class HyCollection extends BaseEntity
     
     /** 总金额 */
     @Excel(name = "总金额")
-    private Long amount;
+    private BigDecimal amount;
 
     /** 是否收款 */
     @Excel(name = "是否收款")
@@ -187,12 +188,12 @@ public class HyCollection extends BaseEntity
     {
         return endTime;
     }
-    public void setAmount(Long amount) 
+    public void setAmount(BigDecimal amount) 
     {
         this.amount = amount;
     }
 
-    public Long getAmount() 
+    public BigDecimal getAmount() 
     {
         return amount;
     }
